requirements.txt
opencv-python
numpy
tqdm

pytorch
torchvision
cudatoolkit
